image-converter/requirements.txt
Pillow>=10.0.0
pillow-avif-plugin>=1.4.2